+++ openrelik_worker_common/archive_utils.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Helper methods for archives."""
+
 import os
 import shutil
 import subprocess
