+++ openrelik_worker_common/debug_utils.py
@@ -11,10 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import os
 
 
-def setup_debugging(port: int = 5678):
+def start_debugger(port: int = 5678):
     """Setup the Python Debugger.
 
     This function initializes the debugpy library, allowing you to attach a debugger to your Python process.
@@ -23,14 +24,11 @@ def setup_debugging(port: int = 5678):
 
     Args:
         port: The port to listen on. Defaults to 5678.
-
-    Returns:
-        None.
     """
     import debugpy
 
-    if os.getenv("OPENRELIK_PYDEBUG") == "1":
-        if os.getenv("OPENRELIK_PYDEBUG_PORT"):
-            port = os.getenv("OPENRELIK_PYDEBUG_PORT")
-        print(f"Starting debugpy on {port}\n")
-        debugpy.listen(("0.0.0.0", int(port)))
+    if os.getenv("OPENRELIK_PYDEBUG_PORT"):
+        port = os.getenv("OPENRELIK_PYDEBUG_PORT")
+
+    print(f"Starting debugpy on {port}\n")
+    debugpy.listen(("0.0.0.0", int(port)))
