+++ tests/test_file_utils.py
@@ -1,10 +1,24 @@
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import unittest
 import unittest.mock
 import os
 import tempfile
 from pathlib import Path
 
-from openrelik_worker_common import utils
+from openrelik_worker_common import file_utils
 
 
 class Utils(unittest.TestCase):
@@ -20,13 +34,8 @@ class Utils(unittest.TestCase):
         if Path(path).resolve().is_file():
             raise AssertionError(f"File exists: {path}")
 
-    def test_dict_to_b64_string(self):
-        """Test dict_to_b64_string function."""
-        result = utils.dict_to_b64_string({"a": "b"})
-        self.assertEqual(result, "eyJhIjogImIifQ==")
-
-    def test_count_lines_in_file(self):
-        """Test count_lines_in_file function."""
+    def test_count_file_lines(self):
+        """Test count_file_lines function."""
         mock_file_content = """Line1
         Line2
         Line3
@@ -34,52 +43,11 @@ class Utils(unittest.TestCase):
 
         file = Path("file.txt")
         file.write_text(mock_file_content, encoding="utf-8")
-        lines = utils.count_lines_in_file(file.absolute())
+        lines = file_utils.count_file_lines(file.absolute())
         self.assertEqual(lines, 3)
         file.unlink()
 
-    def test_get_input_files_input(self):
-        """Test get_input_files function for input_files."""
-        input_file = []
-        input_file.append(utils.create_output_file("test/test"))
-
-        result = utils.get_input_files(pipe_result=None, input_files=input_file)
-        self.assertEqual(result, input_file)
-
-    def test_get_input_files_pipe(self):
-        """Test get_input_files function for pipe_result."""
-        pipe_input = "eyJvdXRwdXRfZmlsZXMiOiAiL3Rlc3QvdGVzdCJ9"
-        result = utils.get_input_files(pipe_result=pipe_input, input_files=None)
-        expected = "/test/test"
-        self.assertEqual(result, expected)
-
-    def test_task_result(self):
-        """˜Test task_result function."""
-        output_files = ["a", "b"]
-        workflow_id = "1234456789"
-        command = "/bin/command"
-        meta = {"key": "value"}
-
-        expected = {
-            "output_files": output_files,
-            "workflow_id": workflow_id,
-            "command": command,
-            "meta": meta,
-            "file_reports": [],
-            "task_report": None,
-        }
-
-        result = utils.task_result(
-            output_files=output_files,
-            workflow_id=workflow_id,
-            command=command,
-            meta=meta,
-            file_reports=[],
-            task_report=None,
-        )
-        self.assertEqual(result, utils.dict_to_b64_string(expected))
-
-    @unittest.mock.patch("openrelik_worker_common.utils.uuid4")
+    @unittest.mock.patch("openrelik_worker_common.file_utils.uuid4")
     def test_create_output_file(self, mock_uuid):
         """Test create_output_file function."""
 
@@ -89,46 +57,48 @@ class Utils(unittest.TestCase):
         mock_uuid.return_value = h
 
         # Test with only an output_path.
-        result = utils.create_output_file(output_base_path="output_path/")
+        result = file_utils.create_output_file(output_base_path="output_path/")
         self.assertEqual(result.display_name, "123456789")
         self.assertEqual(result.path, "output_path/123456789")
 
         # Test with an extra file_name.
-        result = utils.create_output_file(
+        result = file_utils.create_output_file(
             output_base_path="output_path/", display_name="test.txt"
         )
         self.assertEqual(result.display_name, "test.txt")
 
         # Test with an extra file_extension.
-        result = utils.create_output_file(
+        result = file_utils.create_output_file(
             output_base_path="output_path/", display_name="test.txt"
         )
         self.assertEqual(result.display_name, "test.txt")
         self.assertEqual(result.path, "output_path/123456789.txt")
 
         # Test with an extra explicit file_extension.
-        result = utils.create_output_file(
+        result = file_utils.create_output_file(
             output_base_path="output_path/", display_name="test.txt", extension="test"
         )
         self.assertEqual(result.display_name, "test.txt.test")
         self.assertEqual(result.path, "output_path/123456789.test")
 
         # Test with an extra source_file_id OutputFile instance.
-        source_file_id = utils.create_output_file(output_base_path="output_path/")
-        result = utils.create_output_file(
+        source_file_id = file_utils.create_output_file(output_base_path="output_path/")
+        result = file_utils.create_output_file(
             output_base_path="output_path/", source_file_id=source_file_id
         )
         self.assertEqual(result.source_file_id.uuid, source_file_id.uuid)
 
-    @unittest.mock.patch("openrelik_worker_common.utils.uuid4")
+    @unittest.mock.patch("openrelik_worker_common.file_utils.uuid4")
     def test_outputfile_to_dict(self, mock_uuid):
         """Test the outputfile_to_dict function."""
         uuid = unittest.mock.MagicMock()
         uuid.hex = "123456789"
         mock_uuid.return_value = uuid
 
-        parent_outputfile = utils.create_output_file(output_base_path="output_path/")
-        outputfile = utils.create_output_file(
+        parent_outputfile = file_utils.create_output_file(
+            output_base_path="output_path/"
+        )
+        outputfile = file_utils.create_output_file(
             output_base_path="output_path/", source_file_id=parent_outputfile
         )
         result = outputfile.to_dict()
@@ -152,48 +122,48 @@ class Utils(unittest.TestCase):
             ["/etc/../xxx/config", "/xxx/config"],
             ["/etc/../../../../ssh/sshd_config", "/ssh/sshd_config"],
         ]
-        files: utils.OutputFile = []
+        files: file_utils.OutputFile = []
         output_path = tempfile.TemporaryDirectory(delete=False)
         for path in test_paths:
-            file = utils.create_output_file(
+            file = file_utils.create_output_file(
                 output_base_path=output_path.name, original_path=path[0]
             )
             open(file.path, "a", encoding="utf-8").close()
             files.append(file)
 
-        file_tree_root = utils.build_file_tree(output_path.name, files)
+        file_tree_root = file_utils.build_file_tree(output_path.name, files)
 
         for path in test_paths:
             self.assertFileExists(
-                os.path.join(file_tree_root.name, utils.get_path_without_root(path[1]))
+                os.path.join(file_tree_root.name, file_utils.get_relative_path(path[1]))
             )
 
-        utils.delete_file_tree(file_tree_root)
+        file_utils.delete_file_tree(file_tree_root)
 
         # Test with a non OutputFile instance as files array element
-        files = [utils.OutputFile, "not-an-OutputFile"]
-        self.assertIsNone(utils.build_file_tree(output_path, files))
+        files = [file_utils.OutputFile, "not-an-OutputFile"]
+        self.assertIsNone(file_utils.build_file_tree(output_path, files))
 
         # Test with an empty files array
         files = []
-        self.assertIsNone(utils.build_file_tree(output_path, files))
+        self.assertIsNone(file_utils.build_file_tree(output_path, files))
 
     def test_delete_file_tree(self):
         """Test delete_file_tree function."""
         with self.assertRaises(TypeError):
-            utils.delete_file_tree("not-a-temp-directory-object")
+            file_utils.delete_file_tree("not-a-temp-directory-object")
 
         tmpdir = tempfile.TemporaryDirectory(delete=False)
         filepath = os.path.join(tmpdir.name, "testfile")
         open(filepath, "a", encoding="utf-8").close()
 
-        utils.delete_file_tree(tmpdir)
+        file_utils.delete_file_tree(tmpdir)
 
         self.assertFileDoesNotExists(filepath)
 
-    def test_get_path_without_root(self):
-        """Test get_path_without_root function."""
-        relative_path = utils.get_path_without_root("/xxx/yyy/test.txt")
+    def test_get_relative_path(self):
+        """Test get_relative_path function."""
+        relative_path = file_utils.get_relative_path("/xxx/yyy/test.txt")
         self.assertEqual(relative_path, "xxx/yyy/test.txt")
